feat(turboquant): fp16 norm/rnorm storage option (TURBOQUANT_FP16_NORMS)
Norm + rnorm metadata is the largest single overhead beyond the data
bits. For split 3.5-bit it accounts for 32 B / 144 B = 22% of K+V
storage (4 fp32 norms per K side = 16 B; same for V). Cutting these
to fp16 halves the metadata overhead with negligible precision impact:
~0.1% relative error in the dequant rescale, well under the algorithm's
intrinsic Lloyd-Max + QJL noise.

Storage savings (per-token, kv_head, K+V):

  homog b=4 / b=4_r:  144 B  ->  136 B  (3.56x -> 3.76x)
  split  3.5-bit:     144 B  ->  128 B  (3.56x -> 4.00x)
  split  2.25-bit:     96 B  ->   80 B  (estimated)

New stage TURBOQUANT_split_3_5bit_r in eval.sh / throughput.sh /
eval_figure4.sh runs split 3.5-bit with fp16 norms, parallel to the
fp32-norm baseline so the original number can still be reproduced.
The TURBOQUANT_FP16_NORMS=1 env var is independent of TIGHT_PACK and
applies to any TurboQuant config (homog or split).

Implementation:

- Backend allocates norm/rnorm buffers as torch.float16 when the env
  var is set; everything else (idx, qjl_sign, model dtype) unchanged.
- Store kernel casts r_norm before tl.store via
  cache_rnorm_ptr.dtype.element_ty so the same kernel handles both
  dtypes; wrapper casts x_norm similarly. Math stays fp32.
- attend_tc / attend_split_tc append .to(tl.float32) at every
  norm/rnorm tl.load so downstream computation is precision-uniform
  regardless of cache dtype.
- Parity test gains b4_r+fp16-norms and split-3.5_r cases.

+++ vllm/turboquant/attend_split_tc.py
@@ -252,7 +252,7 @@ def _split_attend_kernel(
 
         k_norm_out = tl.load(
             cache_k_norm_out_ptr + meta_addrs, mask=mask_n, other=0.0
-        )
+        ).to(tl.float32)
         k_tile_out_scaled = (
             k_tile_out.to(tl.float32) * k_norm_out[:, None]
         ).to(COMPUTE_DTYPE)
@@ -277,7 +277,7 @@ def _split_attend_kernel(
 
             r_norm_k_out = tl.load(
                 cache_k_rnorm_out_ptr + meta_addrs, mask=mask_n, other=0.0
-            )
+            ).to(tl.float32)
             qjl_dot_k_out = tl.dot(sq_out, tl.trans(qjl_sign_k_out))
             qk_out = qk_out + (
                 qjl_coef_out
@@ -308,7 +308,7 @@ def _split_attend_kernel(
 
         k_norm_reg = tl.load(
             cache_k_norm_reg_ptr + meta_addrs, mask=mask_n, other=0.0
-        )
+        ).to(tl.float32)
         k_tile_reg_scaled = (
             k_tile_reg.to(tl.float32) * k_norm_reg[:, None]
         ).to(COMPUTE_DTYPE)
@@ -333,7 +333,7 @@ def _split_attend_kernel(
 
             r_norm_k_reg = tl.load(
                 cache_k_rnorm_reg_ptr + meta_addrs, mask=mask_n, other=0.0
-            )
+            ).to(tl.float32)
             qjl_dot_k_reg = tl.dot(sq_reg, tl.trans(qjl_sign_k_reg))
             qk_reg = qk_reg + (
                 qjl_coef_reg
@@ -379,7 +379,7 @@ def _split_attend_kernel(
         v_tile_out = tl.load(codebook_out_ptr + v_idx_out)
         v_norm_out = tl.load(
             cache_v_norm_out_ptr + meta_addrs, mask=mask_n, other=0.0
-        )
+        ).to(tl.float32)
         v_tile_out_scaled = (
             v_tile_out.to(tl.float32) * v_norm_out[:, None]
         ).to(COMPUTE_DTYPE)
@@ -398,7 +398,7 @@ def _split_attend_kernel(
             ).to(COMPUTE_DTYPE)
             r_norm_v_out = tl.load(
                 cache_v_rnorm_out_ptr + meta_addrs, mask=mask_n, other=0.0
-            )
+            ).to(tl.float32)
             v_qjl_tile_out = (
                 qjl_sign_v_out.to(tl.float32)
                 * r_norm_v_out[:, None]
@@ -422,7 +422,7 @@ def _split_attend_kernel(
         v_tile_reg = tl.load(codebook_reg_ptr + v_idx_reg)
         v_norm_reg = tl.load(
             cache_v_norm_reg_ptr + meta_addrs, mask=mask_n, other=0.0
-        )
+        ).to(tl.float32)
         v_tile_reg_scaled = (
             v_tile_reg.to(tl.float32) * v_norm_reg[:, None]
         ).to(COMPUTE_DTYPE)
@@ -441,7 +441,7 @@ def _split_attend_kernel(
             ).to(COMPUTE_DTYPE)
             r_norm_v_reg = tl.load(
                 cache_v_rnorm_reg_ptr + meta_addrs, mask=mask_n, other=0.0
-            )
+            ).to(tl.float32)
             v_qjl_tile_reg = (
                 qjl_sign_v_reg.to(tl.float32)
                 * r_norm_v_reg[:, None]
